% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_catch_at_age.R
\name{make_catch_at_age}
\alias{make_catch_at_age}
\title{Make catch at age from catch & bio data}
\usage{
make_catch_at_age(catch, bio)
}
\arguments{
\item{catch}{Catch data frame from stockdbr}

\item{bio}{Bio data frame from stockdbr}
}
\description{
Make catch at age from catch & bio data
}
